requirements.txt
streamlit
python-docx
textstat
language-tool-python
sumy
textblob
nltk